requirements.txt
requests>=2.28.0
pandas>=2.0.0
snowflake-snowpark-python>=1.11.0
snowpipe-streaming>=1.0.0